# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Outfit Builder API
  version: 1.0.0
  description: REST surface of apps/api. The public product catalog is served over
    GraphQL instead (POST /graphql).
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:8080
    description: Local dev
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    TryOnRequest:
      type: object
      properties:
        photo:
          type: string
          pattern: ^data:image\/(jpeg|png|webp);base64,
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
                format: uuid
              category:
                type: string
                enum: *a1
            required:
              - productId
              - category
          minItems: 1
          maxItems: 6
      required:
        - photo
        - items
      additionalProperties: false
    TryOnResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - processing
            - completed
            - failed
        step:
          type: integer
          minimum: 0
        totalSteps:
          type: integer
          minimum: 0
          exclusiveMinimum: true
        resultImageUrl:
          type: string
          nullable: true
          format: uri
        error:
          type: string
          nullable: true
      required:
        - id
        - status
        - step
        - totalSteps
        - resultImageUrl
        - error
...
paths:
...
  /api/try-on:
    post:
      operationId: startTryOn
      summary: Start a virtual try-on — kicks off the first chain step and returns
        immediately
      security: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TryOnRequest"
      responses:
        "201":
          description: Try-on started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TryOnResponse"
        "422":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Upstream try-on service failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Virtual try-on not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/try-on/{id}:
    get:
      operationId: getTryOnStatus
      summary: Poll a try-on request — may itself advance the chain one step further
      security: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TryOnResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...